Readiness outcome integration - session storage

+++ common/javascript/main.js
@@ -29,4 +29,7 @@ document.addEventListener('DOMContentLoaded', function() {
 
 // The expanded state of individual instances of the accordion component persists across page loads using sessionStorage.
 // These will be removed from the session storage so the accordions will collapse rather than stay open.
-sessionStorage.clear();
+// We want to keep session storage on the readiness theme pages to prevent accordions closing on a page refresh 
+if (!window.location.pathname.includes('/transition-readiness-detail/')) {
+  sessionStorage.clear();
+}
